Move dollar parser to separate module
