+++ yash-syntax/src/parser/lex.rs
@@ -45,35 +45,6 @@ pub fn is_token_delimiter_char(c: char) -> bool {
 }
 
 impl Lexer {
-    /// Parses a text unit that starts with `$`.
-    ///
-    /// If the next character is `$`, a parameter expansion, command
-    /// substitution, or arithmetic expansion is parsed. Otherwise, no
-    /// characters are consumed and the return value is `Ok(None)`.
-    pub async fn dollar_unit(&mut self) -> Result<Option<TextUnit>> {
-        let index = self.index();
-        let location = match self.consume_char_if(|c| c == '$').await? {
-            None => return Ok(None),
-            Some(c) => c.location.clone(),
-        };
-
-        // TODO line continuations following $
-        // TODO braced parameter expansion
-        // TODO non-braced parameter expansion
-
-        let location = match self.arithmetic_expansion(location).await? {
-            Ok(result) => return Ok(Some(result)),
-            Err(location) => location,
-        };
-
-        if let Some(result) = self.command_substitution(location).await? {
-            return Ok(Some(result));
-        }
-
-        self.rewind(index);
-        Ok(None)
-    }
-
     /// Parses a backquote unit, possibly preceded by line continuations.
     async fn backquote_unit(
         &mut self,
@@ -407,6 +378,7 @@ impl Lexer {
 // This is here to get better order of Lexer members in the doc.
 mod arith;
 mod command_subst;
+mod dollar;
 mod heredoc;
 pub mod keyword;
 
@@ -416,79 +388,6 @@ mod tests {
     use crate::parser::core::ErrorCause;
     use crate::source::Source;
     use futures::executor::block_on;
-
-    #[test]
-    fn lexer_dollar_unit_no_dollar() {
-        let mut lexer = Lexer::with_source(Source::Unknown, "foo");
-        let result = block_on(lexer.dollar_unit()).unwrap();
-        assert_eq!(result, None);
-
-        let mut lexer = Lexer::with_source(Source::Unknown, "()");
-        let result = block_on(lexer.dollar_unit()).unwrap();
-        assert_eq!(result, None);
-        assert_eq!(block_on(lexer.peek_char()).unwrap().unwrap().value, '(');
-
-        let mut lexer = Lexer::with_source(Source::Unknown, "");
-        let result = block_on(lexer.dollar_unit()).unwrap();
-        assert_eq!(result, None);
-    }
-
-    #[test]
-    fn lexer_dollar_unit_dollar_followed_by_non_special() {
-        let mut lexer = Lexer::with_source(Source::Unknown, "$;");
-        let result = block_on(lexer.dollar_unit()).unwrap();
-        assert_eq!(result, None);
-        assert_eq!(block_on(lexer.peek_char()).unwrap().unwrap().value, '$');
-
-        let mut lexer = Lexer::with_source(Source::Unknown, "$&");
-        let result = block_on(lexer.dollar_unit()).unwrap();
-        assert_eq!(result, None);
-    }
-
-    #[test]
-    fn lexer_dollar_unit_command_substitution() {
-        let mut lexer = Lexer::with_source(Source::Unknown, "$()");
-        let result = block_on(lexer.dollar_unit()).unwrap().unwrap();
-        if let TextUnit::CommandSubst { location, content } = result {
-            assert_eq!(location.line.value, "$()");
-            assert_eq!(location.line.number.get(), 1);
-            assert_eq!(location.line.source, Source::Unknown);
-            assert_eq!(location.column.get(), 1);
-            assert_eq!(content, "");
-        } else {
-            panic!("unexpected result {:?}", result);
-        }
-        assert_eq!(block_on(lexer.peek_char()), Ok(None));
-
-        let mut lexer = Lexer::with_source(Source::Unknown, "$( foo bar )");
-        let result = block_on(lexer.dollar_unit()).unwrap().unwrap();
-        if let TextUnit::CommandSubst { location, content } = result {
-            assert_eq!(location.line.value, "$( foo bar )");
-            assert_eq!(location.line.number.get(), 1);
-            assert_eq!(location.line.source, Source::Unknown);
-            assert_eq!(location.column.get(), 1);
-            assert_eq!(content, " foo bar ");
-        } else {
-            panic!("unexpected result {:?}", result);
-        }
-        assert_eq!(block_on(lexer.peek_char()), Ok(None));
-    }
-
-    #[test]
-    fn lexer_dollar_unit_arithmetic_expansion() {
-        let mut lexer = Lexer::with_source(Source::Unknown, "$((1))");
-        let result = block_on(lexer.dollar_unit()).unwrap().unwrap();
-        if let TextUnit::Arith { content, location } = result {
-            assert_eq!(content, Text(vec![Literal('1')]));
-            assert_eq!(location.line.value, "$((1))");
-            assert_eq!(location.line.number.get(), 1);
-            assert_eq!(location.line.source, Source::Unknown);
-            assert_eq!(location.column.get(), 1);
-        } else {
-            panic!("unexpected result {:?}", result);
-        }
-        assert_eq!(block_on(lexer.peek_char()), Ok(None));
-    }
 
     #[test]
     fn lexer_backquote_not_backquote() {
